Finished Cycle Tracking and some Cleanup
